license_portal/requirements.txt
asgiref==3.7.2
cx-Oracle==8.3.0
Django==3.2
django-cron==0.6.0
djangorestframework==3.14.0
psycopg2==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
